engine/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pydantic>=2.6.0

# Data
pandas>=2.2.0
numpy>=1.26.0

# Exchange
ccxt>=4.2.0

# Config
pyyaml>=6.0.1

# Cache
redis>=5.0.0

# File watching (hot-reload, Sprint 3)
watchdog>=4.0.0

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.3.0
mypy>=1.9.0
